tidying up

+++ art/estimators/classification/pytorch.py
@@ -317,7 +317,7 @@ class PyTorchClassifier(ClassGradientsMixin, ClassifierMixin, PyTorchEstimator):
             )
 
             with torch.no_grad():
-                model_outputs = self._model(torch.from_numpy(x_preprocessed[begin:end]).to(self._device))
+                model_outputs = self._model(torch.from_numpy(x_preprocessed[begin:end]).float().to(self._device))
             output = model_outputs[-1]
             results[begin:end] = output.detach().cpu().numpy()
